--- src/assets/stylesheets/components/all.css
@@ -1,220 +1,0 @@
-.dress {
-  margin-bottom: 50px; }
-
-.dress-name {
-  text-align: center;
-  text-transform: uppercase;
-  margin-top: 10px;
-  margin-bottom: 30px; }
-
-.dress .container-price {
-  display: flex;
-  margin-left: 25px; }
-
-.dress .container-price .price {
-  text-align: right;
-  margin-right: 15px;
-  font-family: 'Baskerville';
-  font-size: 14px; }
-
-.dress .container-price .price .sale-price {
-  font-weight: bold;
-  font-size: 16px;
-  margin-bottom: 3px; }
-
-.button-cart {
-  width: 52px;
-  height: 34px;
-  background-color: #ff9f1c;
-  text-align: center;
-  color: #ffffff;
-  cursor: pointer;
-  border: none;
-  transition: background-color .2s ease-in-out; }
-
-.button-cart:hover {
-  background-color: #e68c10; }
-
-.filter-order {
-  position: absolute;
-  right: 0;
-  top: 0;
-  width: 125px;
-  font-size: 12px;
-  border: 1px solid black;
-  font-family: 'Open Sans', arial; }
-
-.filter-order .options {
-  padding: 10px;
-  cursor: pointer;
-  background-color: #ffffff;
-  display: none; }
-
-.filter-order .options.open {
-  display: block; }
-
-.filter-order .options:not(.label):hover {
-  background-color: #ff9f1c;
-  color: #ffffff; }
-
-.filter-order .options.label {
-  display: block; }
-
-.filter-order .options.label.open {
-  border-bottom: 1px solid black; }
-
-.filter-order .options.label .angle {
-  font-size: 16px;
-  position: absolute;
-  top: 8px;
-  right: 10px; }
-
-@media (max-width: 990px) {
-  .filter-order {
-    right: 20px; } }
-
-@media (max-width: 680px) {
-  .filter-order {
-    display: none; } }
-
-.button-more {
-  margin-top: 10px;
-  margin-bottom: 60px;
-  margin-left: calc(50% - 130px);
-  background-color: #ff9f1c;
-  color: #ffffff;
-  text-transform: uppercase;
-  width: 260px;
-  height: 45px;
-  line-height: 45px;
-  border: none;
-  cursor: pointer;
-  transition: background-color .2s ease-in-out; }
-
-.button-more:hover {
-  background-color: #e68c10; }
-
-.button-more.hide-true {
-  display: none; }
-
-@media (max-width: 680px) {
-  .button-more {
-    width: 100%;
-    margin-left: 0; } }
-
-.box-radio {
-  margin-bottom: 5px;
-  font-family: 'Open Sans', arial;
-  font-size: 14px; }
-
-.radio-buttons {
-  cursor: pointer;
-  position: relative;
-  padding: 2px;
-  border: 1px solid #b0b0b0;
-  -webkit-appearance: none;
-  -moz-appearance: none;
-  appearance: none;
-  margin-right: 10px; }
-
-.radio-buttons:after {
-  content: '';
-  width: 8px;
-  height: 8px;
-  display: block; }
-
-.radio-buttons:checked:after {
-  background-color: #ff9f1c;
-  box-shadow: 0 0 0 transparent !important; }
-
-.content-check {
-  width: 30px;
-  height: 30px;
-  margin-right: 10px;
-  margin-bottom: 10px;
-  position: relative;
-  font-family: 'Open Sans', arial; }
-
-.content-check .label-check {
-  position: absolute;
-  color: #b0b0b0;
-  font-size: 12px;
-  margin: auto;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  -webkit-transform: translate(-50%, -50%);
-  -ms-transform: translate(-50%, -50%); }
-
-.content-check .checks {
-  width: 30px;
-  height: 30px;
-  cursor: pointer;
-  position: relative;
-  border: 1px solid #b0b0b0;
-  -webkit-appearance: none;
-  -moz-appearance: none;
-  appearance: none;
-  z-index: 5; }
-
-.content-check .checks:checked {
-  border: 1px solid #ff9f1c; }
-
-.content-check .checks:checked ~ .label-check {
-  color: #000; }
-
-.box-sizes {
-  width: 160px;
-  height: auto;
-  display: flex;
-  justify-content: space-between;
-  flex-flow: row wrap;
-  align-content: flex-start; }
-
-.box-sizes .flex-null {
-  width: 30px;
-  height: 0;
-  margin-right: 10px; }
-
-.box-sizes.mobile {
-  margin-left: calc(50% - 100px); }
-
-.close-filter {
-  width: 100%;
-  height: 60px;
-  line-height: 60px;
-  padding-left: 20px;
-  padding-right: 20px;
-  display: flex;
-  justify-content: space-between;
-  cursor: pointer;
-  border-bottom: 1px solid #666666; }
-
-.close-filter .label-filters {
-  color: #666666;
-  text-transform: uppercase; }
-
-.close-filter .icons {
-  transform: rotate(45deg);
-  font-weight: 300;
-  font-size: 25px; }
-
-.button-sort-mobile {
-  width: 100%;
-  height: 60px;
-  line-height: 60px;
-  padding-left: 20px;
-  padding-right: 20px;
-  cursor: pointer;
-  position: relative; }
-
-.button-sort-mobile:hover, .button-sort-mobile:active {
-  background-color: #ff9f1c;
-  color: #ffffff; }
-
-.trigger {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%; }
