--- src/assets/stylesheets/components/buttonSortMobile.css
@@ -1,18 +1,0 @@
-.button-sort-mobile {
-  width: 100%;
-  height: 60px;
-  line-height: 60px;
-  padding-left: 20px;
-  padding-right: 20px;
-  cursor: pointer;
-  position: relative; }
-  .button-sort-mobile:hover, .button-sort-mobile:active {
-    background-color: #ff9f1c;
-    color: #ffffff; }
-
-.trigger {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%; }
