+++ src/assets/stylesheets/components/checkboxSizeList.css
@@ -5,11 +5,10 @@
   justify-content: space-between;
   flex-flow: row wrap;
   align-content: flex-start; }
-
-.box-sizes .flex-null {
-  width: 30px;
-  height: 0;
-  margin-right: 10px; }
+  .box-sizes .flex-null {
+    width: 30px;
+    height: 0;
+    margin-right: 10px; }
 
 .box-sizes.mobile {
   margin-left: calc(50% - 100px); }
