+++ src/assets/stylesheets/components/checkboxSizes.css
@@ -5,31 +5,27 @@
   margin-bottom: 10px;
   position: relative;
   font-family: 'Open Sans', arial; }
-
-.content-check .label-check {
-  position: absolute;
-  color: #b0b0b0;
-  font-size: 12px;
-  margin: auto;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  -webkit-transform: translate(-50%, -50%);
-  -ms-transform: translate(-50%, -50%); }
-
-.content-check .checks {
-  width: 30px;
-  height: 30px;
-  cursor: pointer;
-  position: relative;
-  border: 1px solid #b0b0b0;
-  -webkit-appearance: none;
-  -moz-appearance: none;
-  appearance: none;
-  z-index: 5; }
-
-.content-check .checks:checked {
-  border: 1px solid #ff9f1c; }
-
-.content-check .checks:checked ~ .label-check {
-  color: #000; }
+  .content-check .label-check {
+    position: absolute;
+    color: #b0b0b0;
+    font-size: 12px;
+    margin: auto;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    -webkit-transform: translate(-50%, -50%);
+    -ms-transform: translate(-50%, -50%); }
+  .content-check .checks {
+    width: 30px;
+    height: 30px;
+    cursor: pointer;
+    position: relative;
+    border: 1px solid #b0b0b0;
+    -webkit-appearance: none;
+    -moz-appearance: none;
+    appearance: none;
+    z-index: 5; }
+  .content-check .checks:checked {
+    border: 1px solid #ff9f1c; }
+  .content-check .checks:checked ~ .label-check {
+    color: #000; }
