--- src/assets/stylesheets/components/filterSort.css
@@ -1,41 +1,0 @@
-.filter-order {
-  position: absolute;
-  right: 0;
-  top: 0;
-  width: 125px;
-  font-size: 12px;
-  border: 1px solid black;
-  font-family: 'Open Sans', arial; }
-
-.filter-order .options {
-  padding: 10px;
-  cursor: pointer;
-  background-color: #ffffff;
-  display: none; }
-
-.filter-order .options.open {
-  display: block; }
-
-.filter-order .options:not(.label):hover {
-  background-color: #ff9f1c;
-  color: #ffffff; }
-
-.filter-order .options.label {
-  display: block; }
-
-.filter-order .options.label.open {
-  border-bottom: 1px solid black; }
-
-.filter-order .options.label .angle {
-  font-size: 16px;
-  position: absolute;
-  top: 8px;
-  right: 10px; }
-
-@media (max-width: 990px) {
-  .filter-order {
-    right: 20px; } }
-
-@media (max-width: 680px) {
-  .filter-order {
-    display: none; } }
