--- src/assets/stylesheets/components/radioButton.css
@@ -1,24 +1,0 @@
-.box-radio {
-  margin-bottom: 5px;
-  font-family: 'Open Sans', arial;
-  font-size: 14px; }
-
-.radio-buttons {
-  cursor: pointer;
-  position: relative;
-  padding: 2px;
-  border: 1px solid #b0b0b0;
-  -webkit-appearance: none;
-  -moz-appearance: none;
-  appearance: none;
-  margin-right: 10px; }
-
-.radio-buttons:after {
-  content: '';
-  width: 8px;
-  height: 8px;
-  display: block; }
-
-.radio-buttons:checked:after {
-  background-color: #ff9f1c;
-  box-shadow: 0 0 0 transparent !important; }
