+++ src/assets/stylesheets/templates/filterMobile.css
@@ -1,3 +1,13 @@
+.bg-mobile-filter {
+  position: fixed;
+  width: 100%;
+  height: 100%;
+  top: 0;
+  left: 0;
+  background: #ffffff;
+  z-index: 1;
+  display: none; }
+
 .filter-mobile {
   width: 100%;
   min-height: 100%;
@@ -7,49 +17,41 @@
   top: 0;
   left: 0;
   background-color: #ffffff;
-  display: none; }
-
-.filter-mobile .close-filter,
-.filter-mobile .open-close-filters {
-  width: 100%;
-  height: 60px;
-  line-height: 60px;
-  padding-left: 20px;
-  padding-right: 20px;
-  display: flex;
-  justify-content: space-between;
-  cursor: pointer; }
-
-.filter-mobile .close-filter {
-  border-bottom: 1px solid #666666; }
-
-.filter-mobile .close-filter .label-filters {
-  color: #666666; }
-
-.filter-mobile .close-filter .icons {
-  transform: rotate(45deg); }
-
-.filter-mobile .open-close-filters:hover,
-.filter-mobile .open-close-filters:active {
-  background-color: #ff9f1c;
-  color: #ffffff; }
-
-.filter-mobile .label-filters {
-  text-transform: uppercase; }
-
-.filter-mobile .icons {
-  font-weight: 300;
-  font-size: 25px; }
-
-.filter-mobile .content-filters {
-  width: 100%;
-  height: auto;
-  padding: 10px 20px;
-  display: none; }
-
-.filter-mobile .content-filters.open {
-  display: block; }
+  display: none;
+  z-index: 5; }
+  .filter-mobile .close-filter,
+  .filter-mobile .open-close-filters {
+    width: 100%;
+    height: 60px;
+    line-height: 60px;
+    padding-left: 20px;
+    padding-right: 20px;
+    display: flex;
+    justify-content: space-between;
+    cursor: pointer; }
+  .filter-mobile .close-filter {
+    border-bottom: 1px solid #666666; }
+    .filter-mobile .close-filter .label-filters {
+      color: #666666; }
+    .filter-mobile .close-filter .icons {
+      transform: rotate(45deg); }
+  .filter-mobile .open-close-filters:hover {
+    background-color: #ff9f1c;
+    color: #ffffff; }
+  .filter-mobile .label-filters {
+    text-transform: uppercase; }
+  .filter-mobile .icons {
+    font-weight: 300;
+    font-size: 25px; }
+  .filter-mobile .content-filters {
+    width: 100%;
+    height: auto;
+    padding: 10px 20px;
+    display: none; }
+  .filter-mobile .content-filters.open {
+    display: block; }
 
 @media (max-width: 680px) {
-  .filter-mobile.open {
+  .filter-mobile.open,
+  .bg-mobile-filter.open {
     display: block; } }
