--- src/assets/stylesheets/templates/footer.css
@@ -1,7 +1,0 @@
-.footer {
-  width: 100%;
-  height: 35px;
-  line-height: 35px;
-  background-color: #000000;
-  text-align: center;
-  color: #ffffff; }
