--- src/assets/stylesheets/templates/header.css
@@ -1,28 +1,0 @@
-.header {
-  width: 100%;
-  height: 100px;
-  border-bottom: 2px solid #ff9f1c;
-  padding-top: 35px;
-  margin-bottom: 40px; }
-
-.header-content {
-  display: flex;
-  justify-content: space-between;
-  height: 30px; }
-
-.header .bag {
-  position: relative;
-  margin-right: 15px; }
-
-.header .bag .counter {
-  position: absolute;
-  top: 7px;
-  right: -16px;
-  width: 20px;
-  height: 20px;
-  line-height: 20px;
-  background-color: #ff9f1c;
-  border-radius: 50%;
-  color: #ffffff;
-  text-align: center;
-  font-size: 13px; }
