+++ src/templates/filterMobile.jsx
@@ -6,46 +6,49 @@ import ButtonCloseFilterAndSort from '../components/mobile/buttonCloseFilterAndS
 import ButtonFilterMobile from '../components/mobile/buttonFilterMobile';
 
 import RadioButtonColorList from '../components/radioButtonColorList';
-import CheckboxSizeList from '../components/checkboxSizeList';
+import RadioButtonSizeList from '../components/radioButtonSizeList';
+import RadioButtonPriceList from '../components/radioButtonPriceList';
 
 import { openCloseFilter } from '../reducers/filterMobileActions';
 import { openCloseFilterColors } from '../reducers/filterMobileActions';
 import { openCloseFilterSizes } from '../reducers/filterMobileActions';
 import { openCloseFilterPrices } from '../reducers/filterMobileActions';
 
-
-
-
 class FilterMobile extends Component{
   
   render() {
     return(
-      <section className={`filter-mobile ${this.props.openCloseFilterProp}`}>
-        <ButtonCloseFilterAndSort name='Filtrar' onClick={this.props.openCloseFilter}/>
+      <section>
 
-        <ButtonFilterMobile 
-          name='Cores' 
-          onClick={this.props.openCloseFilterColors} 
-          icon={this.props.iconColors} 
-          openClose={this.props.openCloseColors}
-          childComponent={<RadioButtonColorList isMobile='true'/>}
-        />
-        
-        <ButtonFilterMobile 
-          name='Tamanhos' 
-          onClick={this.props.openCloseFilterSizes} 
-          icon={this.props.iconSizes} 
-          openClose={this.props.openCloseSizes}
-          childComponent={<CheckboxSizeList isMobile='true'/>}
-        />
+        <div className={`bg-mobile-filter ${this.props.openCloseFilterProp}`}></div>
 
-        <ButtonFilterMobile 
-          name='Faixa de preços' 
-          onClick={this.props.openCloseFilterPrices} 
-          icon={this.props.iconPrices} 
-          openClose={this.props.openClosePrices}
-          childComponent={<CheckboxSizeList isMobile='true'/>}
-        />
+        <div className={`filter-mobile ${this.props.openCloseFilterProp}`}>
+          <ButtonCloseFilterAndSort name='Filtrar' onClick={this.props.openCloseFilter}/>
+
+          <ButtonFilterMobile 
+            name='Cores' 
+            onClick={this.props.openCloseFilterColors} 
+            icon={this.props.iconColors} 
+            openClose={this.props.openCloseColors}
+            childComponent={<RadioButtonColorList isMobile='true'/>}
+          />
+          
+          <ButtonFilterMobile 
+            name='Tamanhos' 
+            onClick={this.props.openCloseFilterSizes} 
+            icon={this.props.iconSizes} 
+            openClose={this.props.openCloseSizes}
+            childComponent={<RadioButtonSizeList isMobile='true'/>}
+          />
+
+          <ButtonFilterMobile 
+            name='Faixa de preços' 
+            onClick={this.props.openCloseFilterPrices} 
+            icon={this.props.iconPrices} 
+            openClose={this.props.openClosePrices}
+            childComponent={<RadioButtonPriceList />}
+          />
+        </div>
       </section>
     )
   }
